AIForall/Workshop4/Building Intelligent AI agents with Amazon Bedrock and Strands/supply-chain-optimizer/pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "supply-chain-optimizer"
version = "1.0.0"
description = "Multi-agent AI system for supply chain optimization"
requires-python = ">=3.10"
dependencies = [
    "boto3>=1.28.0",
    "botocore>=1.31.0",
    "strands-agents>=1.0.0",
    "python-dotenv>=1.0.0",
    "pydantic>=2.0.0",
    "psycopg2-binary>=2.9.0",
    "sqlalchemy>=2.0.0",
    "aws-xray-sdk>=2.12.0",
    "pydantic-settings>=2.0.0",
    "flask>=3.0.0",
    "flask-cors>=4.0.0",
    "flask-jwt-extended>=4.5.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-cov>=4.1.0",
    "pytest-asyncio>=0.21.0",
    "hypothesis>=6.88.0",
    "black>=23.0.0",
    "isort>=5.12.0",
    "flake8>=6.1.0",
    "mypy>=1.7.0",
    "python-dotenv>=1.0.0",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "--cov=src --cov-report=html --cov-report=term-missing"

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false

[tool.black]
line-length = 100
target-version = ['py310']

[tool.isort]
profile = "black"
line_length = 100
